feat: backport profiles and settings
